(#28) Some updates
Updated composer to put into dev and changed the assertion.

I also ran some server changes to get the tests to work but it's not acting as I expect
exactly. With this level of uncertainty, I am going to hold off employing this until I
get a better handle on it. For now, will continue using PhpUnit and Behat.