(#32) Added more files and contexts for Behat
Updated the behat.sh file and behat.yml.

Also added some new contexts for saving screenshots, etc.

This branch is currently problematic b/c it is not using the right database. It is using the dev
database which is not what I want. Will continue to look at it but then again, may just go with
phpunit acceptance testing if easier to implement.